FIX paths to REST endpoints and more
If we want to specify our REST endpoints URLS as absolute ones, we need
to include the application context. That is, the root under which the
application will be deployed to Tomcat. Otherwise, images or paths won't
be found.

You can see the errors for that in the console (developer tools in the
browser).